requirements.txt
langchain>=0.1.0
langchain-google-genai
langchain-community>=0.0.20
python-dotenv>=1.0.0